Refactor selectors for static properties.
