+++ src/html/component-editor/properties-editor/static/static-essential-condition/static-essential-condition.less
@@ -20,7 +20,6 @@
 }
 
 .value-symbol {
-  .uk-width-1-2;
   .uk-flex;
   .uk-flex-row;
   .uk-flex-center;
