+++ src/html/component-editor/properties-editor/static/static-essential-condition/static-reg-essential-condition.ts
@@ -7,10 +7,10 @@ import {
 } from '../../../../util/data-interfaces'
 import { getEssentialityText, getNextEssentiality } from '../../../../util/utilities'
 import { map } from 'lit/directives/map.js'
-import StaticSelectorsProperty from '../static-selectors-property'
+import StaticRegSelectorsProperty from '../static-reg-selectors-property'
 
 @customElement('static-reg-essential-condition')
-export default class StaticRegEssentialCondition extends StaticSelectorsProperty {
+export default class StaticRegEssentialCondition extends StaticRegSelectorsProperty {
   static styles = css`${unsafeCSS(style_less)}`
   @property() declare property: IVariableRegulatorEssentialStaticProperty
 
@@ -41,23 +41,23 @@ export default class StaticRegEssentialCondition extends StaticSelectorsProperty
       <div class="property-body">
         ${this.renderNameplate()}
         <div class="value-section">
-          <div class="value-symbol gap">
-            <select id="input-selector" class="uk-select" @change="${this.inputChanged}"
-                    ?disabled="${this.property.target === null}">
+          <div class="value-symbol uk-width-3-5 gap">
+            <select id="input-selector" class="uk-select" @change="${this.inputChanged}">
               <option value="${null}">---</option>
               ${map(this.getInputSelectorItems(), (item) => html`
                 <option value="${item}">${item}</option>
               `)}
             </select>
             <span>${this.getEssentialitySymbol()}</span>
-            <select id="target-selector" class="uk-select" @change="${this.targetChanged}">
+            <select id="target-selector" class="uk-select" @change="${this.targetChanged}"
+                    ?disabled="${this.property.input === null}">
               <option value="${null}">---</option>
               ${map(this.getTargetSelectorItems(), (item) => html`
                 <option value="${item}">${item}</option>
               `)}
             </select>
           </div>
-          <div class="value-symbol" @click="${() => {
+          <div class="value-symbol uk-width-2-5" @click="${() => {
             this.toggleEssentiality()
           }}">
             <span class="essentiality">
