+++ src/html/component-editor/properties-editor/static/static-essential/static-essential.less
@@ -17,7 +17,6 @@
 }
 
 .value-symbol {
-  .uk-width-1-2;
   .uk-flex;
   .uk-flex-row;
   .uk-flex-center;
