+++ src/html/component-editor/properties-editor/static/static-essential/static-fn-essential.ts
@@ -15,13 +15,13 @@ export default class StaticFnEssential extends abstractStaticProperty {
       <div class="property-body">
         ${this.renderNameplate(false)}
         <div class="value-section">
-          <div class="value-symbol gap">
+          <div class="value-symbol uk-width-3-5 gap">
             <label for="target">Fn: </label>
             <div id="target" class="uk-margin-small-right uk-text-bold">${this.property.target}</div>
             <label for="input">Input: </label>
             <div id="input" class="uk-margin-small-right uk-text-bold">${this.property.input}</div>
           </div>
-          <div class="value-symbol">
+          <div class="value-symbol uk-width-2-5">
             <span class="essentiality">
               ${getEssentialityText(this.property.value)}
             </span>
