+++ src/html/component-editor/properties-editor/static/static-essential/static-reg-essential.ts
@@ -26,12 +26,12 @@ export default class StaticRegEssential extends abstractStaticProperty {
       <div class="property-body">
         ${this.renderNameplate(false)}
         <div class="value-section">
-          <div class="value-symbol">
+          <div class="value-symbol uk-width-3-5">
             <div class="uk-margin-small-right uk-text-bold">${this.property.input}</div>
             <div class="uk-margin-small-right">${this.getEssentialitySymbol()}</div>
             <div class="uk-margin-small-right uk-text-bold">${this.property.target}</div>
           </div>
-          <div class="value-symbol">
+          <div class="value-symbol uk-width-2-5">
             <span class="essentiality">
               ${getEssentialityText(this.property.value)}
             </span>
