+++ src/html/component-editor/properties-editor/static/static-monotonic-condition/static-monotonic-condition.less
@@ -36,7 +36,6 @@
 }
 
 .value-symbol {
-  .uk-width-1-2;
   .uk-flex;
   .uk-flex-row;
   .uk-flex-center;
