+++ src/html/component-editor/properties-editor/static/static-monotonic/static-monotonic.less
@@ -33,7 +33,6 @@
 }
 
 .value-symbol {
-  .uk-width-1-2;
   .uk-flex;
   .uk-flex-row;
   .uk-flex-center;
